[frontend] Rename profile to settings (#101)
