+++ frontend/app/settings/change-password/page.tsx
@@ -1,4 +1,4 @@
-import ChangePasswordForm from "@/app/ui/profile/change-password-form";
+import ChangePasswordForm from "@/app/ui/settings/change-password-form";
 
 export default function ChangePasswordPage() {
   return <ChangePasswordForm />;
