+++ frontend/app/settings/display/page.tsx
