+++ frontend/app/settings/page.tsx
@@ -1,4 +1,4 @@
-import ProfileForm from "@/app/ui/profile/profile-form";
+import ProfileForm from "@/app/ui/settings/profile-form";
 
 export default function ProfilePage() {
   return <ProfileForm />;
